app/image_predictor/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
filelock==3.13.3
fsspec==2024.3.1
gunicorn==21.2.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
torch==2.2.1
torchvision==0.17.1
typing_extensions==4.10.0
urllib3==2.2.1
